+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 1.4.0
+## Added
+- documenteditor/config/editorconfig/Customization#mobile
+- documenteditor/config/editorconfig/Customization#pointerMode
+- documenteditor/config/editorconfig/Customization#slidePlayerBackground
+- documenteditor/config/editorconfig/Customization#wordHeadingsColor
+- support hwp, hwpx, pages, numbers, key formats
+- sq-AL empty file templates
+
+## Changed
+- documenteditor/config/editorconfig/Customization#submitForm type boolean or SubmitForm
+- deprecated documenteditor/config/editorconfig/Customization#mobileForceView
+
 ## 1.3.0
 ## Added
 - documenteditor/config/editorconfig/customization/Logo#imageLight